Minor fixes

+++ BERT/networks.py
@@ -262,12 +262,12 @@ class LightningBertPretrainedClassifier(LightningModule):
             total_labels.append(x["labels"])
             total_unique_ids.append(x["unique_ids"])
             total_prediction_probs.append(x["prediction_probs"])
-        avg_loss = torch.stack(total_loss).mean()
-        unique_ids = torch.cat(total_unique_ids)
-        predictions = torch.cat(total_predictions)
-        prediction_probs = torch.cat(total_prediction_probs, dim=0)
-        labels = torch.cat(total_labels)
-        correct = predictions.eq(labels.view_as(predictions))
+        avg_loss = torch.stack(total_loss).double().mean()
+        unique_ids = torch.cat(total_unique_ids).long()
+        predictions = torch.cat(total_predictions).long()
+        prediction_probs = torch.cat(total_prediction_probs, dim=0).double()
+        labels = torch.cat(total_labels).long()
+        correct = predictions.eq(labels.view_as(predictions)).long()
         accuracy = correct.double().mean()
         save_predictions(self.hparams.output_path,
                          unique_ids.data.cpu().numpy(),
